+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# x.y.z
+
+- Add boolean type widgets
+
 # 0.4.14
 
 - Fix localizing config when config has a table and a field with choices/elements